refactor: SlackClient 인터페이스로 분리
- AcceptanceTest에서 MethodsClient 모킹 제거
- AcceptanceTest에서 테스트용 SlackClient 주입
