+++ backend/src/test/java/com/pickpick/acceptance/AcceptanceTest.java
@@ -2,7 +2,8 @@ package com.pickpick.acceptance;
 
 import com.pickpick.auth.support.JwtTokenProvider;
 import com.pickpick.support.DatabaseCleaner;
-import com.slack.api.methods.MethodsClient;
+import com.pickpick.support.SlackClient;
+import com.pickpick.support.TestConfig;
 import io.restassured.RestAssured;
 import java.time.Clock;
 import org.junit.jupiter.api.AfterEach;
@@ -11,10 +12,11 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.context.annotation.Import;
 
+@Import(value = TestConfig.class)
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public class AcceptanceTest {
@@ -25,8 +27,8 @@ public class AcceptanceTest {
     @SpyBean
     protected Clock clock;
 
-    @MockBean
-    protected MethodsClient slackClient;
+    @Autowired
+    protected SlackClient slackClient;
 
     @Autowired
     protected JwtTokenProvider jwtTokenProvider;
