+++ backend/src/test/java/com/pickpick/acceptance/channel/ChannelAcceptanceTest.java
@@ -40,7 +40,7 @@ public class ChannelAcceptanceTest extends AcceptanceTest {
 
     private void 채널_목록_생성(final String memberSlackId) {
         for (Channel channel : ChannelFixture.createAllChannels()) {
-            채널_생성(memberSlackId, channel, slackClient);
+            채널_생성(memberSlackId, channel);
         }
     }
 }
