+++ backend/src/test/java/com/pickpick/acceptance/channel/ChannelSubscriptionAcceptanceTest.java
@@ -46,7 +46,7 @@ class ChannelSubscriptionAcceptanceTest extends AcceptanceTest {
     @Test
     void 채널_구독() {
         // given
-        채널_생성(MEMBER_SLACK_ID, NOTICE.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, NOTICE.create());
         long channelId = 1L;
 
         // when
@@ -60,7 +60,7 @@ class ChannelSubscriptionAcceptanceTest extends AcceptanceTest {
     @Test
     void 채널_구독_취소() {
         // given
-        채널_생성(MEMBER_SLACK_ID, NOTICE.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, NOTICE.create());
         long channelId = 1L;
 
         구독_요청(token, channelId);
@@ -76,11 +76,11 @@ class ChannelSubscriptionAcceptanceTest extends AcceptanceTest {
     @Test
     void 채널_구독_조회() {
         // given
-        채널_생성(MEMBER_SLACK_ID, NOTICE.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, NOTICE.create());
         long noticeId = 1L;
         구독_요청(token, noticeId);
 
-        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create());
         long freeChatId = 2L;
         구독_요청(token, freeChatId);
 
@@ -95,11 +95,11 @@ class ChannelSubscriptionAcceptanceTest extends AcceptanceTest {
     @Test
     void 구독_채널_순서_변경() {
         // given
-        채널_생성(MEMBER_SLACK_ID, NOTICE.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, NOTICE.create());
         long noticeId = 1L;
         구독_요청(token, noticeId);
 
-        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create());
         long freeChatId = 2L;
         구독_요청(token, freeChatId);
 
@@ -117,11 +117,11 @@ class ChannelSubscriptionAcceptanceTest extends AcceptanceTest {
     @ValueSource(ints = {0, -1})
     void 구독_채널_순서_변경_시_1보다_작은_순서가_들어올_경우_예외_발생(int invalidViewOrder) {
         // given
-        채널_생성(MEMBER_SLACK_ID, NOTICE.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, NOTICE.create());
         long noticeId = 1L;
         구독_요청(token, noticeId);
 
-        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create());
         long freeChatId = 2L;
         구독_요청(token, freeChatId);
 
@@ -141,11 +141,11 @@ class ChannelSubscriptionAcceptanceTest extends AcceptanceTest {
     @Test
     void 구독_채널_순서_변경_시_중복된_순서가_들어올_경우_예외_발생() {
         // given
-        채널_생성(MEMBER_SLACK_ID, NOTICE.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, NOTICE.create());
         long noticeId = 1L;
         구독_요청(token, noticeId);
 
-        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create());
         long freeChatId = 2L;
         구독_요청(token, freeChatId);
 
@@ -165,10 +165,10 @@ class ChannelSubscriptionAcceptanceTest extends AcceptanceTest {
     @Test
     void 구독_채널_순서_변경_시_해당_멤버가_구독한_적_없는_채널_ID가_포함된_경우_예외_발생() {
         // given
-        채널_생성(MEMBER_SLACK_ID, NOTICE.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, NOTICE.create());
         long noticeId = 1L;
 
-        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create());
         long freeChatId = 2L;
         구독_요청(token, freeChatId);
 
@@ -188,11 +188,11 @@ class ChannelSubscriptionAcceptanceTest extends AcceptanceTest {
     @Test
     void 구독_채널_순서_변경_시_해당_멤버의_모든_구독_채널이_요청에_포함되지_않을_경우_예외_발생() {
         // given
-        채널_생성(MEMBER_SLACK_ID, NOTICE.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, NOTICE.create());
         long noticeId = 1L;
         구독_요청(token, noticeId);
 
-        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create());
         long freeChatId = 2L;
         구독_요청(token, freeChatId);
 
@@ -211,7 +211,7 @@ class ChannelSubscriptionAcceptanceTest extends AcceptanceTest {
     @Test
     void 구독_중인_채널_다시_구독_요청() {
         // given
-        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create());
         long freeChatId = 1L;
         구독_요청(token, freeChatId);
 
@@ -226,7 +226,7 @@ class ChannelSubscriptionAcceptanceTest extends AcceptanceTest {
     @Test
     void 구독하지_않은_채널_구독_취소() {
         // given
-        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create(), slackClient);
+        채널_생성(MEMBER_SLACK_ID, FREE_CHAT.create());
         long freeChatId = 1L;
 
         // when
