+++ backend/src/test/java/com/pickpick/support/TestConfig.java
@@ -1,0 +1,15 @@
+package com.pickpick.support;
+
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Primary;
+
+@TestConfiguration
+public class TestConfig {
+
+    @Primary
+    @Bean
+    public SlackClient slackClient() {
+        return new SlackClientTest();
+    }
+}
